requirements.txt
flask
openai
python-dotenv
pypdf
